pyproject.toml
[project]
name = "iamoney"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cachetools>=6.2.2",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "openai>=2.9.0",
    "pillow>=12.0.0",
    "pyngrok>=7.5.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
